Fixed compiler warning -Wextra

+++ sdk/libgdb/gdb-cortex-m.c
@@ -688,7 +688,7 @@ int thread_info(unsigned int gdb_thread_id, char * buf)
 			if (thread_id != THINKOS_THREAD_IDLE) {
 				int irq;
 				for (irq = 0; irq < THINKOS_IRQ_MAX; ++irq) {
-					if (thinkos_rt.irq_th[irq] == thread_id) {
+					if (thinkos_rt.irq_th[irq] == (int)thread_id) {
 						break;
 					}
 				}
@@ -926,9 +926,9 @@ int target_file_read(const char * name, char * dst,
 					  unsigned int offs, unsigned int size)
 {
 	char * src;
-	int len;
-	int cnt;
-	int i;
+	unsigned int len;
+	unsigned int cnt;
+	unsigned int i;
 
 	if (prefix(name, "target.xml")) {
 		src = (char *)target_xml;
